requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confluent-kafka==2.0.2
constantly==15.1.0
cryptography==39.0.1
Django==4.1.6
django-celery-beat==2.4.0
django-redis==4.12.1
django-redis-cache==3.0.0
django-timezone-field==5.0
djangorestframework==3.14.0
Faker==16.6.1
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.4
priority==1.3.0
prompt-toolkit==3.0.37
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
redis==4.5.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
